fix(scotia): sprated the files into 3 .py files
